requirements.txt
# Core dependencies
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.21.0

# Model and training
timm>=0.9.0
scikit-learn>=1.0.0

# Data processing
pyarrow>=10.0.0
fastparquet>=0.8.0
pandas>=1.5.0
Pillow>=9.0.0
requests>=2.28.0

# Configuration
PyYAML>=6.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Computer Vision
opencv-python>=4.6.0

# Optional dependencies for advanced features
# wandb>=0.12.0  # For experiment tracking (uncomment if needed)
# umap-learn>=0.5.0  # For UMAP visualization (uncomment if needed)
# tensorboard>=2.10.0  # For TensorBoard logging (uncomment if needed)

# Development dependencies (optional)
# pytest>=6.0.0
# black>=21.0.0
# flake8>=3.9.0